Move classes to a correct directory
